+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "nanoid": "^3.2.0",
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-datepicker": "^4.6.0",
